move zzfx stuff to separate file
