+++ packages/superdough/index.mjs
@@ -8,4 +8,5 @@ export * from './superdough.mjs';
 export * from './sampler.mjs';
 export * from './helpers.mjs';
 export * from './synth.mjs';
+export * from './zzfx.mjs';
 export * from './logger.mjs';
